BISERVER-9820 - disabled32.png and enabled32.png from unknown source
Removed the icon with copyright issues, enabled 32.png.  Moved the accept.png from data access over into its place.  It is 24 px, so I renamed it enabled24.png and updated the reference in plugin.xml.  Also, updated plugin.xml so that the test perspective sample works as intended.

Also, while in there, got with Wes and replaced the lock icon in case it was a copyright issue as well.