[MONDRIAN-1914] MongOLAP not logging Agg Pipeline queries sent to MongoDB
This change sets the location of log4j.xml in a system property in OSGIBoot.  Also set is the log4j.ignoreTCL=true property, which governs whether it starts with the
classloader associated with the thread when loading classes.  This could cause CCEs, since the OSGI classloader should be used.