BISERVER-13000 As developer I want to be able to assemble CE pentaho-platform for java 8
Java 8 adds new methods to abstract classes used in doclet api. For unit
testing we use to have hierarchy to this classes as a stubs. For java 7
they compiled successfully, for java 8 is not since the lack of
implementations of new methods.

For unit testing we can just have mocks. We need to add mockito as ivy
dependency and rewrite testing code. After this project will be able to be
assembled using java 8 jdk.