[SP-2513] Backport of BISERVER-11649 - Mondrian role mappers don't pass the role name of a user when his username and group are both 'Admin' (5.4 Suite)
- UserRoleDaoResource
Track every update of roles, and if this update is touching a user in current session, updating IPentahoSession.SESSION_ROLES attribute, that represents the roles, avalible for current user.
We are doing it via userRoleDaoService, because data there is always up-to-date
- Added few helper methods and changed visibility of some methods for testing reasongs

-UserRoleDaoResource_RoleUpdatesTest
A separate test, written specially for checking, that roles were/were not (depending on whether user is session-user or not) updated.

-UserRoleDaoResourceTest
Changed few tests a little,(added several mocks), as method content changed.