[BACKLOG-18798] De-hydrating the pentaho-server of worker-nodes config (#3812)
* [BACKLOG-18798] De-hydrating the pentaho-server of worker-nodes config:

In order to remove the need for wiring the IActionInvoker bean within the pentahoObjects.spring.xml file, we need to have the ability to instantiate the detault action invoker directly within the ActionAdapterQuartzJob class. With the previous code base, we did not have this ability, because the DefaultActionInvoker class resided within the extensions module, which is downstream from the scheduler module (scheduler) where the class needed to be referenced. Therefore some refactoring/moving of the default action invoker and other related classes needed to happen.

As a result, the concept of a 'LocalActionInvoker' was introduced. The DefaultActionInvoker class still exists, but has been moved from extensions to scheduler module, so that it can be accessed directly from within ActionAdapterQuartzJob. It has been modified slightly to fetch the instance of streamProviver directly from the params map via the 'QuartzScheduler.RESERVEDMAPKEY_STREAMPROVIDER' key (this is how ActionAdapterQuartzJob fetched the streamProvider prior to the action-invocation refactoring). LocalActionInvoker within extensions (formerly known as DefaultActionInvoker) extends from the new DefaultActionInvoker within scheduler and overrides the getStreamProvider method to fetch it in a more generic fashion that allows us to serialize the parameters and invoke the action remotely. All worker node enabled code (spring wiring, ActionInvokerDelgator) should now directly access the LocalActionInvoker in extensions, the DefaultActionInvoker in scheduling should only be used from within ActionAdapterQuartzJob

* [BACKLOG-18798] De-hydrating the pentaho-server of worker-nodes config:

In order to remove the need for wiring the IActionInvoker bean within the pentahoObjects.spring.xml file, we need to have the ability to instantiate the detault action invoker directly within the ActionAdapterQuartzJob class. With the previous code base, we did not have this ability, because the DefaultActionInvoker class resided within the extensions module, which is downstream from the scheduler module (scheduler) where the class needed to be referenced. Therefore some refactoring/moving of the default action invoker and other related classes needed to happen.

As a result, the concept of a 'LocalActionInvoker' was introduced. The DefaultActionInvoker class still exists, but has been moved from extensions to scheduler module, so that it can be accessed directly from within ActionAdapterQuartzJob. It has been modified slightly to fetch the instance of streamProviver directly from the params map via the 'QuartzScheduler.RESERVEDMAPKEY_STREAMPROVIDER' key (this is how ActionAdapterQuartzJob fetched the streamProvider prior to the action-invocation refactoring). LocalActionInvoker within extensions (formerly known as DefaultActionInvoker) extends from the new DefaultActionInvoker within scheduler and overrides the getStreamProvider method to fetch it in a more generic fashion that allows us to serialize the parameters and invoke the action remotely. All worker node enabled code (spring wiring, ActionInvokerDelgator) should now directly access the LocalActionInvoker in extensions, the DefaultActionInvoker in scheduling should only be used from within ActionAdapterQuartzJob

* [BACKLOG-18798] De-hydrating the pentaho-server of worker-nodes config:

In order to remove the need for wiring the IActionInvoker bean within the pentahoObjects.spring.xml file, we need to have the ability to instantiate the detault action invoker directly within the ActionAdapterQuartzJob class. With the previous code base, we did not have this ability, because the DefaultActionInvoker class resided within the extensions module, which is downstream from the scheduler module (scheduler) where the class needed to be referenced. Therefore some refactoring/moving of the default action invoker and other related classes needed to happen.

As a result, the concept of a 'LocalActionInvoker' was introduced. The DefaultActionInvoker class still exists, but has been moved from extensions to scheduler module, so that it can be accessed directly from within ActionAdapterQuartzJob. It has been modified slightly to fetch the instance of streamProviver directly from the params map via the 'QuartzScheduler.RESERVEDMAPKEY_STREAMPROVIDER' key (this is how ActionAdapterQuartzJob fetched the streamProvider prior to the action-invocation refactoring). LocalActionInvoker within extensions (formerly known as DefaultActionInvoker) extends from the new DefaultActionInvoker within scheduler and overrides the getStreamProvider method to fetch it in a more generic fashion that allows us to serialize the parameters and invoke the action remotely. All worker node enabled code (spring wiring, ActionInvokerDelgator) should now directly access the LocalActionInvoker in extensions, the DefaultActionInvoker in scheduling should only be used from within ActionAdapterQuartzJob