BACKLOG-20622
Error when opening Buyer Report.  Replace the Hyperlink function on product code.   This was using the Repository (Legacy) (4.8) function.  The new repository (jackrabbit) started in 5.0.   This should have been corrected in the report.  Back in the  5.0 & 6.0 release, I believe we may have accomodated this.   This sample report should have updated the hyperlink fuction on product code when the report was updated in 7.0.