[BISERVER-14398] Pen server Perf Degradation when importing objects
Cause of the issue was identified as a gradually increasing number of orphaned
JCR sessions.  This occurred because our caching strategy would skip logout
for certain sessions to allow them to be reused, but never did a logout
on eviction from the cache.

This change logs out on eviction when a session has no active use.  In general
at the time of eviction _all_ sessions should be ready for logout, given 5 minutes
of inactivity, but if cache settings are more restrictive, or if activity is
very high, this will prevent killing an in-use session.

https://jira.pentaho.com/browse/BISERVER-14398